requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
google-generative-ai
duckduckgo-search
